# src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"TSOAAddress": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"landmark": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"houseNumber": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"pinCode": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"isDefault": {
						"type": "boolean"
					},
					"deleted": {
						"type": "boolean"
					},
					"tag": {
						"type": "string"
					},
					"user": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"mobileNumber",
					"email",
					"landmark",
					"address",
					"houseNumber",
					"city",
					"pinCode",
					"state",
					"date",
					"isDefault",
					"deleted",
					"tag",
					"user",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/TSOAAddress"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddressesResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"docs": {
								"items": {
									"$ref": "#/components/schemas/TSOAAddress"
								},
								"type": "array"
							}
						},
						"required": [
							"count",
							"docs"
						],
						"type": "object"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"data",
					"statusCode",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserRole": {
				"enum": [
					"SUPER_ADMIN",
					"ADMIN",
					"USER",
					"VENDOR"
				],
				"type": "string"
			},
...
			"TSOALabel": {
				"properties": {
					"Size": {
						"type": "string"
					},
					"Material": {
						"type": "string"
					},
					"Color": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"TSOAVariant": {
				"properties": {
					"label": {
						"$ref": "#/components/schemas/TSOALabel"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"comparePrice": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"SKU": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"price",
					"comparePrice",
					"quantity",
					"SKU",
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOAVariantOptionType": {
				"enum": [
					"Color",
					"Material",
					"Size"
				],
				"type": "string"
			},
			"TSOAVariantOption": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/TSOAVariantOptionType"
					},
					"values": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"values",
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOALocation": {
				"properties": {
					"type": {
						"type": "string"
					},
					"coordinates": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"type",
					"coordinates"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOAddress": {
				"properties": {
					"location": {
						"$ref": "#/components/schemas/TSOALocation"
					},
					"pinCode": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"state": {
						"type": "string"
					}
				},
				"required": [
					"location",
					"pinCode",
					"city",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOAVendorInfo": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"businessEmail": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"gstNumber": {
						"type": "string"
					},
					"panNumber": {
						"type": "string"
					},
					"addressProof": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "boolean"
					},
					"commissionPercent": {
						"type": "number",
						"format": "double"
					},
					"deleted": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					},
					"profileImage": {
						"type": "string"
					},
					"date": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"businessName",
					"businessEmail",
					"mobileNumber",
					"gstNumber",
					"panNumber",
					"addressProof",
					"status",
					"commissionPercent",
					"deleted",
					"createdAt",
					"updatedAt",
					"__v",
					"profileImage",
					"date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOAVendor": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"address": {
						"$ref": "#/components/schemas/TSOAddress"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"vendorStatus": {
						"type": "boolean"
					},
					"isMobileVerified": {
						"type": "boolean"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"deleted": {
						"type": "boolean"
					},
					"vendorInfo": {
						"$ref": "#/components/schemas/TSOAVendorInfo"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					},
					"expirationTime": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"varificationToken": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"isVendorProfileComplete": {
						"type": "boolean"
					},
					"isApproved": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"address",
					"email",
					"mobileNumber",
					"role",
					"vendorStatus",
					"isMobileVerified",
					"isEmailVerified",
					"deleted",
					"vendorInfo",
					"createdAt",
					"updatedAt",
					"__v",
					"expirationTime",
					"varificationToken",
					"isVendorProfileComplete",
					"isApproved",
					"date",
					"name",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOAProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"brand": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"quantityDependent": {
						"type": "boolean"
					},
					"isApproved": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"isFeatured": {
						"type": "boolean"
					},
					"metaDescription": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"ratingCount": {
						"type": "number",
						"format": "double"
					},
					"viewCount": {
						"type": "number",
						"format": "double"
					},
					"deleted": {
						"type": "boolean"
					},
					"variants": {
						"items": {
							"$ref": "#/components/schemas/TSOAVariant"
						},
						"type": "array"
					},
					"variantOptions": {
						"items": {
							"$ref": "#/components/schemas/TSOAVariantOption"
						},
						"type": "array"
					},
					"vendor": {
						"$ref": "#/components/schemas/TSOAVendor"
					},
					"userRole": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"description",
					"brand",
					"images",
					"type",
					"status",
					"quantityDependent",
					"isApproved",
					"date",
					"isFeatured",
					"metaDescription",
					"rating",
					"ratingCount",
					"viewCount",
					"deleted",
					"variants",
					"variantOptions",
					"vendor",
					"userRole",
					"createdBy",
					"createdAt",
					"updatedAt",
					"slug",
					"__v"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApprovalStatus": {
				"enum": [
					"PENDING",
					"APPROVED",
					"DISAPPROVED"
				],
				"type": "string"
			},
			"TSOAUser": {
				"properties": {
					"address": {
						"$ref": "#/components/schemas/TSOAAddress"
					},
					"email": {
						"type": "string"
					},
					"mobileNumber": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"isVendorProfileComplete": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"vendorStatus": {
						"type": "boolean"
					},
					"isApproved": {
						"$ref": "#/components/schemas/ApprovalStatus"
					},
					"isMobileVerified": {
						"type": "boolean"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"date": {
						"type": "string"
					},
					"deleted": {
						"type": "boolean"
					},
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"varificationToken": {
						"type": "string",
						"nullable": true
					},
					"expirationTime": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"address",
					"email",
					"mobileNumber",
					"role",
					"isVendorProfileComplete",
					"isActive",
					"vendorStatus",
					"isApproved",
					"isMobileVerified",
					"isEmailVerified",
					"date",
					"deleted",
					"_id",
					"name",
					"varificationToken",
					"expirationTime",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TSOACart": {
				"properties": {
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"_id": {
						"type": "string"
					},
					"product": {
						"$ref": "#/components/schemas/TSOAProduct"
					},
					"vendor": {
						"type": "string"
					},
					"variant": {
						"$ref": "#/components/schemas/TSOAVariant"
					},
					"user": {
						"$ref": "#/components/schemas/TSOAUser"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"__v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"quantity",
					"_id",
					"product",
					"vendor",
					"variant",
					"user",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Add Bearer prefix on jwt token"
			}
		}
	},
	"info": {
		"title": "Market Place Apis Documentation",
		"version": "1.0.0",
		"description": "Market Place Node Backend Apis and Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"email": "marketplace@gmail.com",
			"name": "marketplace",
			"url": "http://marketplace.in"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/address": {
			"post": {
				"operationId": "AddAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressResponse"
								}
							}
						}
					}
				},
				"summary": "Add Address",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isDefault": {
										"type": "boolean"
									},
									"address": {
										"type": "string"
									},
									"pinCode": {
										"type": "string"
									},
									"city": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"firstName": {
										"type": "string"
									},
									"mobileNumber": {
										"type": "string"
									}
								},
								"required": [
									"isDefault",
									"address",
									"pinCode",
									"city",
									"lastName",
									"firstName",
									"mobileNumber"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAddresses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressesResponse"
								}
							}
						}
					}
				},
				"summary": "fetch Addresses",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/address/{addressId}": {
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressResponse"
								}
							}
						}
					}
				},
				"summary": "Update Address",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "addressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isDefault": {
										"type": "boolean"
									},
									"address": {
										"type": "string"
									},
									"pinCode": {
										"type": "string"
									},
									"city": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"firstName": {
										"type": "string"
									},
									"mobileNumber": {
										"type": "string"
									}
								},
								"required": [
									"isDefault",
									"address",
									"pinCode",
									"city",
									"lastName",
									"firstName",
									"mobileNumber"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressResponse"
								}
							}
						}
					}
				},
				"summary": "fetch Address By id",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "addressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressResponse"
								}
							}
						}
					}
				},
				"summary": "Delete Address",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "addressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/address/default/{addressId}": {
			"put": {
				"operationId": "MakeDefaultAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AddressResponse"
								}
							}
						}
					}
				},
				"summary": "make default Address",
				"tags": [
					"Address"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "addressId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/cart": {
			"post": {
				"operationId": "AddItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Add Item on cart",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"quantity": {
										"type": "number",
										"format": "double"
									},
									"variantId": {
										"type": "string"
									},
									"productId": {
										"type": "string"
									}
								},
								"required": [
									"quantity",
									"variantId",
									"productId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetCartInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TSOACart"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "get cart info",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/cart/bulk-products": {
			"post": {
				"operationId": "AddBulkItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "Add Bulk Item on cart",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"items": {
										"items": {
											"properties": {
												"quantity": {
													"type": "number",
													"format": "double"
												},
												"variantId": {
													"type": "string"
												},
												"productId": {
													"type": "string"
												}
											},
											"required": [
												"quantity",
												"variantId",
												"productId"
											],
											"type": "object"
										},
										"type": "array"
									}
								},
								"required": [
									"items"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/cart/decrease/{itemId}": {
			"put": {
				"operationId": "DecreaseItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "decrease item quantity",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cart/increase/{itemId}": {
			"put": {
				"operationId": "IncreaseItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "increase item quantity",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cart/{itemId}": {
			"delete": {
				"operationId": "RemoveItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "remove item from cart",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cart/clear": {
			"get": {
				"operationId": "ClearCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"summary": "clear cart",
				"tags": [
					"Cart"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}